fix: restructuring of all modules (putting into one file)
